backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-dotenv
pydantic
reportlab
requests
pandas
numpy
scikit-learn
geopy
email-validator